# admin-backend/docs/swagger.yaml
definitions:
  admin_proto.AdminApiItem:
    properties:
      createdAt:
        type: string
      describe:
        type: string
      id:
        type: integer
      isEnabled:
        type: boolean
      key:
        type: string
      name:
        type: string
      path:
        type: string
      permissionId:
        type: integer
      updatedAt:
        type: string
    type: object
...
  admin_proto.AdminPermissionInfo:
    properties:
      apis:
        description: 接口列表
        items:
          $ref: '#/definitions/admin_proto.AdminApiItem'
        type: array
      createdAt:
        type: string
      describe:
        description: 权限描述
        type: string
      id:
        description: 权限ID
        type: integer
      isEnabled:
        description: 是否启用
        type: boolean
      key:
        description: 权限唯一标识符
        type: string
      menuId:
        description: 权限对应的菜单ID
        type: integer
      menuName:
        description: 菜单名称
        type: string
      menuPath:
        description: 菜单路由
        type: string
      name:
        description: 权限名称
        type: string
      redirect:
        description: 重定向地址
        type: string
      type:
        description: 权限类型
        type: string
      typeText:
        description: 权限类型
        type: string
      updatedAt:
        type: string
    type: object
...
  admin_proto.PermissionListItem:
    properties:
      apis:
        description: 接口列表
        items:
          $ref: '#/definitions/admin_proto.AdminApiItem'
        type: array
      createdAt:
        type: string
      describe:
        description: 权限描述
        type: string
      id:
        description: 权限ID
        type: integer
      isEnabled:
        description: 是否启用
        type: boolean
      key:
        description: 权限唯一标识符
        type: string
      menuId:
        description: 权限对应的菜单ID
        type: integer
      menuName:
        description: 菜单名称
        type: string
      menuPath:
        description: 菜单路由
        type: string
      name:
        description: 权限名称
        type: string
      type:
        description: 权限类型
        type: string
      typeText:
        description: 权限类型
        type: string
      updatedAt:
        type: string
    type: object
...
  admin_proto.RespAdminPermissionListData:
    properties:
      list:
        items:
          $ref: '#/definitions/admin_proto.PermissionListItem'
        type: array
      total:
        type: integer
    type: object
...
  code.Message:
    properties:
      code:
        allOf:
        - $ref: '#/definitions/code_proto.ErrorCode'
        description: 错误码
      data: {}
      msg:
        description: 错误码描述
        type: string
      reason:
        description: 错误具体信息
        type: string
    type: object
  code_proto.ErrorCode:
    enum:
    - 0
    - 1
    - 100001
    - 100002
    - 100003
    - 200001
    - 200002
    - 200003
    - 200004
    - 200005
    - 300001
    - 300002
    - 300003
    - 400001
    - 400002
    - 400003
    - 500001
    - 500002
    - 500003
    - 500004
    - 600001
    - 600002
    - 600003
    - 600004
    - 600005
    - 700001
    - 800001
    type: integer
    x-enum-comments:
      ErrorCode_AdminAccountInvalid: 账号已被封禁
      ErrorCode_AdminAccountNotExist: 账号不存在或已被删除
      ErrorCode_AdminAccountPasswordInvalid: 账号或密码错误
      ErrorCode_AdminApiKeyExist: 接口键名已存在
      ErrorCode_AdminApiNameExist: 接口名称已存在
      ErrorCode_AdminApiNotExist: 接口不存在货已失效
      ErrorCode_AdminApiPathExist: 接口路径已存在
      ErrorCode_AdminMenuNotExist: 菜单不存在
      ErrorCode_AdminPermissionExist: 该菜单下权限已存在
      ErrorCode_AdminPermissionKeyExist: 权限键名已存在
      ErrorCode_AdminPermissionKeyNeed: 权限键名不能为空
      ErrorCode_AdminPermissionNameNeed: 权限名称不能为空
      ErrorCode_AdminPermissionTypeInvalid: 权限类型无效
      ErrorCode_AdminRoleNotExist: 角色不存在
      ErrorCode_AuthTokenFailed: 鉴权失败
      ErrorCode_AuthTokenForbidden: 没有访问权限
      ErrorCode_AuthTokenInfoInvalid: Token信息无效
      ErrorCode_AuthTokenInspectInvalid: BearerToken无效
      ErrorCode_AuthTokenInvalid: BearerToken无效
      ErrorCode_Error: 失败
      ErrorCode_ReadContextRequestBodyFailed: 获取请求体内容失败
      ErrorCode_RecordNValidCanNotDeleted: 生效中的记录不能删除
      ErrorCode_RecordNotExist: 记录不存在
      ErrorCode_RequestBodyInvalid: 请求体参数无效
      ErrorCode_RequestParamsInvalid: 请求参数无效
      ErrorCode_RequestQueryInvalid: 请求查询参数无效
      ErrorCode_Success: 成功
    x-enum-varnames:
    - ErrorCode_Success
    - ErrorCode_Error
    - ErrorCode_ReadContextRequestBodyFailed
    - ErrorCode_RecordNotExist
    - ErrorCode_RecordNValidCanNotDeleted
    - ErrorCode_AuthTokenFailed
    - ErrorCode_AuthTokenInvalid
    - ErrorCode_AuthTokenInspectInvalid
    - ErrorCode_AuthTokenInfoInvalid
    - ErrorCode_AuthTokenForbidden
    - ErrorCode_RequestBodyInvalid
    - ErrorCode_RequestQueryInvalid
    - ErrorCode_RequestParamsInvalid
    - ErrorCode_AdminAccountPasswordInvalid
    - ErrorCode_AdminAccountNotExist
    - ErrorCode_AdminAccountInvalid
    - ErrorCode_AdminApiNameExist
    - ErrorCode_AdminApiPathExist
    - ErrorCode_AdminApiKeyExist
    - ErrorCode_AdminApiNotExist
    - ErrorCode_AdminPermissionKeyExist
    - ErrorCode_AdminPermissionExist
    - ErrorCode_AdminPermissionKeyNeed
    - ErrorCode_AdminPermissionNameNeed
    - ErrorCode_AdminPermissionTypeInvalid
    - ErrorCode_AdminMenuNotExist
    - ErrorCode_AdminRoleNotExist
info:
  contact: {}
paths:
...
  /admin/permission/add:
    post:
      consumes:
      - application/json
      description: 添加权限
      parameters:
      - description: "@gotags:\tlabel:\"权限描述\" binding:\"required,trimBlank\""
        in: query
        name: describe
        required: true
        type: string
      - description: "@gotags:\tlabel:\"是否启用\""
        in: query
        name: enabled
        type: boolean
      - description: "@gotags:\tlabel:\"权限ID\" binding:\"required,min=1\""
        in: query
        minimum: 1
        name: id
        required: true
        type: integer
      - description: "@gotags:\tlabel:\"权限唯一标识符\" binding:\"required,trimBlank\""
        in: query
        name: key
        required: true
        type: string
      - description: "@gotags:\tlabel:\"权限对应的菜单ID\" binding:\"required,min=1\""
        in: query
        minimum: 1
        name: menuId
        required: true
        type: integer
      - description: "@gotags:\tlabel:\"权限名称\" binding:\"required,trimBlank\""
        in: query
        name: name
        required: true
        type: string
      - description: "@gotags:\tlabel:\"重定向地址\""
        in: query
        name: redirect
        type: string
      - description: "@gotags:\tlabel:\"权限类型\" binding:\"required,trimBlank\""
        in: query
        name: type
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: desc
          schema:
            allOf:
            - $ref: '#/definitions/code.Message'
            - properties:
                data:
                  type: object
              type: object
      summary: 添加权限
      tags:
      - 权限接口相关
  /admin/permission/add_menu_permissions:
    post:
      consumes:
      - application/json
      description: 指定菜单创建查看编辑删除权限
      parameters:
      - description: "@gotags:\tlabel:\"菜单ID\" binding:\"required,min=1\""
        in: query
        minimum: 1
        name: menuId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: desc
          schema:
            allOf:
            - $ref: '#/definitions/code.Message'
            - properties:
                data:
                  type: object
              type: object
      summary: 指定菜单创建查看编辑删除权限
      tags:
      - 权限接口相关
  /admin/permission/bind_apis:
    post:
      consumes:
      - application/json
      description: 绑定权限接口
      parameters:
      - collectionFormat: csv
        description: "@gotags:\tlabel:\"接口ID列表\" binding:\"required,min=1\""
        in: query
        items:
          type: integer
        minItems: 1
        name: apiIds
        required: true
        type: array
      - description: "@gotags:\tlabel:\"权限ID\" binding:\"required,min=1\""
        in: query
        minimum: 1
        name: permissionId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: desc
          schema:
            allOf:
            - $ref: '#/definitions/code.Message'
            - properties:
                data:
                  type: object
              type: object
      summary: 绑定权限接口
      tags:
      - 权限接口相关
  /admin/permission/delete:
    post:
      consumes:
      - application/json
      description: 删除权限
      parameters:
      - description: "@gotags:\tlabel:\"权限ID\" binding:\"required,min=1\""
        in: query
        minimum: 1
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: desc
          schema:
            allOf:
            - $ref: '#/definitions/code.Message'
            - properties:
                data:
                  type: object
              type: object
      summary: 删除权限
      tags:
      - 权限接口相关
  /admin/permission/edit:
    post:
      consumes:
      - application/json
      description: 编辑权限
      parameters:
      - description: "@gotags:\tlabel:\"权限描述\" binding:\"required,trimBlank\""
        in: query
        name: describe
        required: true
        type: string
      - description: "@gotags:\tlabel:\"启用状态\""
        in: query
        name: enabled
        type: boolean
      - description: "@gotags:\tlabel:\"权限ID\" binding:\"required,min=1\""
        in: query
        minimum: 1
        name: id
        required: true
        type: integer
      - description: "@gotags:\tlabel:\"权限唯一标识符\" binding:\"required,trimBlank\""
        in: query
        name: key
        required: true
        type: string
      - description: "@gotags:\tlabel:\"权限对应的菜单ID\" binding:\"required,min=1\""
        in: query
        minimum: 1
        name: menuId
        required: true
        type: integer
      - description: "@gotags:\tlabel:\"权限名称\" binding:\"required,trimBlank\""
        in: query
        name: name
        required: true
        type: string
      - description: "@gotags:\tlabel:\"重定向地址\""
        in: query
        name: redirect
        type: string
      - description: "@gotags:\tlabel:\"权限类型\" binding:\"required,trimBlank\""
        in: query
        name: type
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: desc
          schema:
            allOf:
            - $ref: '#/definitions/code.Message'
            - properties:
                data:
                  type: object
              type: object
      summary: 编辑权限
      tags:
      - 权限接口相关
  /admin/permission/enable:
    post:
      consumes:
      - application/json
      description: 权限启用|禁用
      parameters:
      - description: "@gotags:\tlabel:\"权限ID\" binding:\"required,min=1\""
        in: query
        minimum: 1
        name: id
        required: true
        type: integer
      - description: "@gotags:\tlabel:\"启用状态\""
        in: query
        name: isEnabled
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: desc
          schema:
            allOf:
            - $ref: '#/definitions/code.Message'
            - properties:
                data:
                  type: object
              type: object
      summary: 权限启用|禁用
      tags:
      - 权限接口相关
  /admin/permission/info:
    post:
      consumes:
      - application/json
      description: 权限详情
      parameters:
      - description: "@gotags:\tlabel:\"权限ID\" binding:\"required,min=1\""
        in: query
        minimum: 1
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: desc
          schema:
            allOf:
            - $ref: '#/definitions/code.Message'
            - properties:
                data:
                  $ref: '#/definitions/admin_proto.AdminPermissionInfo'
              type: object
      summary: 权限详情
      tags:
      - 权限接口相关
  /admin/permission/list:
    post:
      consumes:
      - application/json
      description: 权限列表
      parameters:
      - description: 权限唯一标识符
        in: query
        name: key
        type: string
      - description: 菜单ID
        in: query
        name: menuId
        type: integer
      - description: 权限名称
        in: query
        name: name
        type: string
      - description: 权限类型
        in: query
        name: type
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: desc
          schema:
            allOf:
            - $ref: '#/definitions/code.Message'
            - properties:
                data:
                  $ref: '#/definitions/admin_proto.RespAdminPermissionListData'
              type: object
      summary: 权限列表
      tags:
      - 权限接口相关
  /admin/permission/unbind_api:
    post:
      consumes:
      - application/json
      description: 解绑权限接口
      parameters:
      - description: "@gotags:\tlabel:\"接口ID\" binding:\"required,min=1\""
        in: query
        minimum: 1
        name: apiId
        required: true
        type: integer
      - description: "@gotags:\tlabel:\"权限ID\" binding:\"required,min=1\""
        in: query
        minimum: 1
        name: permissionId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: desc
          schema:
            allOf:
            - $ref: '#/definitions/code.Message'
            - properties:
                data:
                  type: object
              type: object
      summary: 解绑权限接口
      tags:
      - 权限接口相关
...
swagger: "2.0"